fix keys as symbols for price history

+++ lib/schwab_rb/clients/base_client.rb
@@ -892,7 +892,7 @@ module SchwabRb
     def format_date_as_millis(var_name, dt)
       assert_type(var_name, dt, [Date, DateTime])
       dt = DateTime.new(dt.year, dt.month, dt.day) unless dt.is_a?(DateTime)
-      (dt.to_time.to_f * 1000).to_i
+      (dt.to_time.to_i * 1000).to_i
     end
 
     def normalize_start_and_end_datetimes(start_datetime, end_datetime)
